+++ app/src/main/java/cwnuchrome/aac_cwnu_it_2015_1/ActionMultiWord.java
@@ -378,7 +378,7 @@ public abstract class ActionMultiWord extends ActionItem {
             @NonNull HashMap<Long, QueryWordInfo> queryMap,
             @NonNull HashMap<Long, Double> eval_map,
             long entire_collection_count,
-            long average_document_length
+            double average_document_length
     ) {
 
         for (Map.Entry<Long, QueryWordInfo> entry : queryMap.entrySet()) {
